Expose isolate as a function and refactor logger
