+++ src/helpers/detect-package-manager.ts
@@ -2,8 +2,7 @@ import fs from "fs-extra";
 import assert from "node:assert";
 import { execSync } from "node:child_process";
 import path from "node:path";
-import { createLogger, readTypedJsonSync } from "~/utils";
-import { getConfig } from "./config";
+import { readTypedJsonSync, useLogger } from "~/utils";
 import type { PackageManifest } from "./create-packages-registry";
 import { getLockfileFileName } from "./process-lockfile";
 
@@ -35,7 +34,7 @@ export function detectPackageManager(workspaceRoot: string): PackageManager {
 }
 
 function inferFromManifest(workspaceRoot: string) {
-  const log = createLogger(getConfig().logLevel);
+  const log = useLogger();
 
   const rootManifest = readTypedJsonSync<PackageManifest>(
     path.join(workspaceRoot, "package.json")
