+++ src/utils/logger.ts
@@ -1,34 +1,68 @@
 import chalk from "chalk";
-import type { IsolateConfigResolved } from "~/helpers/config";
-
+import type { IsolateConfigResolved, LogLevel } from "~/helpers/config";
+/**
+ * The Logger defines an interface that can be used to pass in a different
+ * logger object in order to intercept all the logging output. We keep the
+ * handlers separate from the logger object itself, so that we can change the
+ * handlers but do not bother the user with having to handle logLevel.
+ */
 export type Logger = {
-  debug(...args: any[]): void;
-  info(...args: any[]): void;
-  warn(...args: any[]): void;
-  error(...args: any[]): void;
+  debug(...args: unknown[]): void;
+  info(...args: unknown[]): void;
+  warn(...args: unknown[]): void;
+  error(...args: unknown[]): void;
 };
 
-export function createLogger(
+let _loggerHandlers: Logger = {
+  debug(...args: unknown[]) {
+    console.log(chalk.blue("debug"), ...args);
+  },
+  info(...args: unknown[]) {
+    console.log(chalk.green("info"), ...args);
+  },
+  warn(...args: unknown[]) {
+    console.log(chalk.yellow("warning"), ...args);
+  },
+  error(...args: unknown[]) {
+    console.log(chalk.red("error"), ...args);
+  },
+};
+
+const _logger: Logger = {
+  debug(...args: unknown[]) {
+    if (_logLevel === "debug") {
+      _loggerHandlers.debug(args);
+    }
+  },
+  info(...args: unknown[]) {
+    if (_logLevel === "debug" || _logLevel === "info") {
+      _loggerHandlers.info(args);
+    }
+  },
+  warn(...args: unknown[]) {
+    if (_logLevel === "debug" || _logLevel === "info" || _logLevel === "warn") {
+      _loggerHandlers.warn(args);
+    }
+  },
+  error(...args: unknown[]) {
+    _loggerHandlers.error(args);
+  },
+};
+
+let _logLevel: LogLevel = "info";
+
+export function setLogger(logger: Logger) {
+  _loggerHandlers = logger;
+  return _logger;
+}
+
+export function setLogLevel(
   logLevel: IsolateConfigResolved["logLevel"]
 ): Logger {
-  return {
-    debug(...args: any[]) {
-      if (logLevel === "debug") {
-        console.log(chalk.blue("debug"), ...args);
-      }
-    },
-    info(...args: any[]) {
-      if (logLevel === "debug" || logLevel === "info") {
-        console.log(chalk.green("info"), ...args);
-      }
-    },
-    warn(...args: any[]) {
-      if (logLevel === "debug" || logLevel === "info" || logLevel === "warn") {
-        console.log(chalk.yellow("warning"), ...args);
-      }
-    },
-    error(...args: any[]) {
-      console.log(chalk.red("error"), ...args);
-    },
-  };
+  _logLevel = logLevel;
+  return _logger;
+}
+
+export function useLogger() {
+  return _logger;
 }
